Move to react-scripts, resolves audit issues
