+++ .babelrc
@@ -1,0 +1,16 @@
+{
+    "presets": [
+        "@babel/react",
+        "@babel/typescript",
+        [
+            "@babel/env",
+            {
+                "modules": "commonjs"
+            }
+        ],
+    ],
+    "plugins": [
+        "@babel/proposal-class-properties",
+        "@babel/proposal-object-rest-spread"
+    ]
+}
