--- example/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "historical-dates-ui-example",
-  "homepage": "https://UUDigitalHumanitieslab.github.io/historical-dates-ui",
-  "version": "0.0.0",
-  "license": "MIT",
-  "private": true,
-  "dependencies": {
-    "historical-dates-ui": "file:..",
-    "prop-types": "^15.6.2",
-    "react": "^16.4.1",
-    "react-dom": "^16.4.1",
-    "react-scripts": "^1.1.4"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
-  }
-}
