--- rollup.config.js
@@ -1,41 +1,0 @@
-import typescript from 'rollup-plugin-typescript2'
-import commonjs from 'rollup-plugin-commonjs'
-import external from 'rollup-plugin-peer-deps-external'
-import postcss from 'rollup-plugin-postcss-modules'
-import resolve from 'rollup-plugin-node-resolve'
-import url from 'rollup-plugin-url'
-
-import pkg from './package.json'
-
-export default {
-    input: 'src/index.tsx',
-    output: [{
-            file: pkg.main,
-            format: 'cjs',
-            sourcemap: true
-        },
-        {
-            file: pkg.module,
-            format: 'es',
-            sourcemap: true
-        }
-    ],
-    plugins: [
-        external(),
-        postcss({
-            modules: true,
-            writeDefinitions: true
-        }),
-        url(),
-        resolve(),
-        typescript({
-            rollupCommonJSResolveHack: true
-        }),
-        commonjs({
-            include: ['node_modules/historical-dates/dist/historical-dates.js'],
-            namedExports: {
-                'node_modules/historical-dates/dist/historical-dates.js': ['calcEaster', 'createDate', 'createDateFromString', 'fromRomanNumber', 'toRomanNumber', 'InvalidDateException', 'HistoricalDate', 'RomanDate', 'RomanDays', 'RomanTexts', 'RomanMonths']
-            }
-        })
-    ]
-}
